227 Truncated alembic revision history
