--- alembic/versions/18bc72fce61_add_version_to_download_events.py
@@ -1,22 +1,0 @@
-"""Add version to download events
-
-Revision ID: 18bc72fce61
-Revises: 46ea35b1ee8
-Create Date: 2014-06-25 15:17:20.977825
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '18bc72fce61'
-down_revision = '46ea35b1ee8'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
