--- alembic/versions/1ccdaebac05_add_database_tracked_game_versions.py
@@ -1,26 +1,0 @@
-"""Add database-tracked game versions
-
-Revision ID: 1ccdaebac05
-Revises: 3b5f1c92dda
-Create Date: 2014-07-06 21:02:32.478886
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '1ccdaebac05'
-down_revision = '3b5f1c92dda'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
