--- alembic/versions/1d46e8d4483_add_mod_versioning.py
@@ -1,28 +1,0 @@
-"""Add mod versioning
-
-Revision ID: 1d46e8d4483
-Revises: 2650a2191fe
-Create Date: 2014-06-10 01:29:49.567535
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '1d46e8d4483'
-down_revision = '2650a2191fe'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'ksp_version')
-    op.drop_column('mod', 'keywords')
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('keywords', sa.VARCHAR(length=256), autoincrement=False, nullable=True))
-    op.add_column('mod', sa.Column('ksp_version', sa.VARCHAR(length=16), autoincrement=False, nullable=True))
-    ### end Alembic commands ###
