--- alembic/versions/1ef399cbd2b_add_downloads_and_followers_to_mod.py
@@ -1,28 +1,0 @@
-"""Add downloads and followers to Mod
-
-Revision ID: 1ef399cbd2b
-Revises: 29cdccab86f
-Create Date: 2014-06-11 17:10:26.480478
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '1ef399cbd2b'
-down_revision = '29cdccab86f'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('download_count', sa.Integer(), server_default='0', nullable=False))
-    op.add_column('mod', sa.Column('follower_count', sa.Integer(), server_default='0', nullable=False))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'follower_count')
-    op.drop_column('mod', 'download_count')
-    ### end Alembic commands ###
