--- alembic/versions/1fb18596264_increase_length_of_backgroundmedia_field.py
@@ -1,26 +1,0 @@
-"""Increase length of backgroundMedia field
-
-Revision ID: 1fb18596264
-Revises: 35e7157127b
-Create Date: 2014-06-01 22:18:01.642532
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '1fb18596264'
-down_revision = '35e7157127b'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
