--- alembic/versions/20025e7424d_add_indexs_and_adjust_forum_accounts.py
@@ -1,34 +1,0 @@
-"""Add indexs and adjust forum accounts
-
-Revision ID: 20025e7424d
-Revises: 59abfc7b45d
-Create Date: 2014-06-10 19:05:40.599138
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '20025e7424d'
-down_revision = '59abfc7b45d'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.create_index(op.f('ix_mod_description'), 'mod', ['description'], unique=False)
-    op.create_index(op.f('ix_mod_name'), 'mod', ['name'], unique=False)
-    op.add_column('user', sa.Column('forumId', sa.Integer(), nullable=True))
-    op.create_index(op.f('ix_user_email'), 'user', ['email'], unique=False)
-    op.create_index(op.f('ix_user_username'), 'user', ['username'], unique=False)
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_user_username'), table_name='user')
-    op.drop_index(op.f('ix_user_email'), table_name='user')
-    op.drop_column('user', 'forumId')
-    op.drop_index(op.f('ix_mod_name'), table_name='mod')
-    op.drop_index(op.f('ix_mod_description'), table_name='mod')
-    ### end Alembic commands ###
