+++ alembic/versions/2019_08_29-00_40_42-6ffd5dd5efab.py
@@ -1,0 +1,20 @@
+"""Initial revision (truncated)
+Previous revisions were deleted at Thu Aug 29 00:40:42
+
+Revision ID: 6ffd5dd5efab
+Revises: None
+Create Date: 2016-04-08 22:58:14.178434
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = '6ffd5dd5efab'
+down_revision = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass
