--- alembic/versions/217af36c820_add_password_reset_expiry.py
@@ -1,26 +1,0 @@
-"""Add password reset expiry
-
-Revision ID: 217af36c820
-Revises: f9897e968a
-Create Date: 2014-06-28 11:04:23.134422
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '217af36c820'
-down_revision = 'f9897e968a'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('passwordResetExpiry', sa.DateTime(), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'passwordResetExpiry')
-    ### end Alembic commands ###
