--- alembic/versions/263bcce1dc0_add_blog_posts.py
@@ -1,22 +1,0 @@
-"""Add blog posts
-
-Revision ID: 263bcce1dc0
-Revises: 5330eb08753
-Create Date: 2014-06-24 19:57:17.971590
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '263bcce1dc0'
-down_revision = '5330eb08753'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
