--- alembic/versions/2650a2191fe_add_source_link.py
@@ -1,26 +1,0 @@
-"""Add source link
-
-Revision ID: 2650a2191fe
-Revises: 3a3487af1b8
-Create Date: 2014-06-08 22:17:58.499334
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '2650a2191fe'
-down_revision = '3a3487af1b8'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('source_link', sa.String(length=256), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'source_link')
-    ### end Alembic commands ###
