--- alembic/versions/29344aa34d9_create_auth_table.py
@@ -1,26 +1,0 @@
-"""create auth table
-
-Revision ID: 29344aa34d9
-Revises: 474b0fb0868
-Create Date: 2016-02-07 02:25:58.668610
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '29344aa34d9'
-down_revision = '474b0fb0868'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
