--- alembic/versions/299a128fba2_add_more_fine_version_control_schema.py
@@ -1,27 +1,0 @@
-"""Add more fine version control schema
-
-Revision ID: 299a128fba2
-Revises: 1ccdaebac05
-Create Date: 2014-07-17 21:26:07.956319
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '299a128fba2'
-down_revision = '1ccdaebac05'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('default_version_id', sa.Integer(), nullable=True))
-    op.add_column('modversion', sa.Column('sort_index', sa.Integer(), nullable=True))
-    ### end Alembic commands ###
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('modversion', 'sort_index')
-    op.drop_column('mod', 'default_version_id')
-    ### end Alembic commands ###
