--- alembic/versions/29cdccab86f_add_updated_time_to_mod.py
@@ -1,25 +1,0 @@
-"""Add updated time to Mod
-
-Revision ID: 29cdccab86f
-Revises: 2bee80c31e1
-Create Date: 2014-06-11 15:13:25.434598
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '29cdccab86f'
-down_revision = '2bee80c31e1'
-
-from alembic import op
-import sqlalchemy as sa
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('updated', sa.DateTime(), nullable=True))
-    ### end Alembic commands ###
-    
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'updated')
-    ### end Alembic commands ###
