--- alembic/versions/2ff737052f3_wire_up_download_events.py
@@ -1,22 +1,0 @@
-"""Wire up download events
-
-Revision ID: 2ff737052f3
-Revises: e591232705
-Create Date: 2014-06-25 14:47:30.675276
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '2ff737052f3'
-down_revision = 'e591232705'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
