--- alembic/versions/35597d56e8d_add_ckan_boolean_to_mod_table.py
@@ -1,26 +1,0 @@
-"""Add ckan boolean to mod table
-
-Revision ID: 35597d56e8d
-Revises: 18af22fa9e4
-Create Date: 2014-12-12 20:11:22.250080
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '35597d56e8d'
-down_revision = '50b5a95300c'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('ckan', sa.Boolean(), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'ckan')
-    ### end Alembic commands ###
