--- alembic/versions/35e7157127b_initial_database_commit.py
@@ -1,26 +1,0 @@
-"""Initial database commit
-
-Revision ID: 35e7157127b
-Revises: None
-Create Date: 2014-06-01 21:15:45.151351
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '35e7157127b'
-down_revision = None
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
