--- alembic/versions/3afc6d6bfd9_remove_index_from_descriptions.py
@@ -1,26 +1,0 @@
-"""Remove index from descriptions
-
-Revision ID: 3afc6d6bfd9
-Revises: 299a128fba2
-Create Date: 2014-08-14 19:38:32.006037
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '3afc6d6bfd9'
-down_revision = '299a128fba2'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('ix_mod_description', table_name='mod')
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.create_index('ix_mod_description', 'mod', ['description'], unique=False)
-    ### end Alembic commands ###
