--- alembic/versions/3c9f83ee88c_add_mod_short_description_and_remove_.py
@@ -1,28 +1,0 @@
-"""Add mod short_description and remove installation
-
-Revision ID: 3c9f83ee88c
-Revises: 20025e7424d
-Create Date: 2014-06-10 23:56:48.617107
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '3c9f83ee88c'
-down_revision = '20025e7424d'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('short_description', sa.Unicode(length=1000), nullable=True))
-    op.drop_column('mod', 'installation')
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('installation', sa.VARCHAR(length=100000), autoincrement=False, nullable=True))
-    op.drop_column('mod', 'short_description')
-    ### end Alembic commands ###
