--- alembic/versions/3d4136d1ae1_expand_size_of_background_field_again.py
@@ -1,26 +1,0 @@
-"""Expand size of background field again
-
-Revision ID: 3d4136d1ae1
-Revises: e6fcc34862
-Create Date: 2015-02-08 14:34:52.207688
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '3d4136d1ae1'
-down_revision = 'e6fcc34862'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
