--- alembic/versions/3f35f3fa395_extend_length_of_changelogs.py
@@ -1,22 +1,0 @@
-"""Extend length of changelogs
-
-Revision ID: 3f35f3fa395
-Revises: 18bc72fce61
-Create Date: 2014-06-27 00:26:05.310245
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '3f35f3fa395'
-down_revision = '18bc72fce61'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
