--- alembic/versions/46d0c3d3104_update_mod_media_schema.py
@@ -1,26 +1,0 @@
-"""Update mod/media schema
-
-Revision ID: 46d0c3d3104
-Revises: 2ef141f5132
-Create Date: 2014-06-06 00:34:39.866314
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '46d0c3d3104'
-down_revision = '2ef141f5132'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('background', sa.String(length=32), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'background')
-    ### end Alembic commands ###
