--- alembic/versions/46ea35b1ee8_fix_error_in_download_events.py
@@ -1,22 +1,0 @@
-"""Fix error in download events
-
-Revision ID: 46ea35b1ee8
-Revises: 2ff737052f3
-Create Date: 2014-06-25 14:59:01.317728
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '46ea35b1ee8'
-down_revision = '2ff737052f3'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
