--- alembic/versions/474b0fb0868_add_dark_theme.py
@@ -1,26 +1,0 @@
-"""Add dark theme
-
-Revision ID: 474b0fb0868
-Revises: 3d4136d1ae1
-Create Date: 2015-08-30 15:54:20.885280
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '474b0fb0868'
-down_revision = '3d4136d1ae1'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('dark_theme', sa.Boolean(), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'dark_theme')
-    ### end Alembic commands ###
