--- alembic/versions/4ea6bc2045a_add_timestamp_to_mod_versions.py
@@ -1,26 +1,0 @@
-"""Add timestamp to mod versions
-
-Revision ID: 4ea6bc2045a
-Revises: 1d46e8d4483
-Create Date: 2014-06-10 01:38:00.449409
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '4ea6bc2045a'
-down_revision = '1d46e8d4483'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('modversion', sa.Column('created', sa.DateTime(), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('modversion', 'created')
-    ### end Alembic commands ###
