--- alembic/versions/4f528e69e89_add_schema_for_followers_and_download_.py
@@ -1,22 +1,0 @@
-"""Add schema for followers and download count
-
-Revision ID: 4f528e69e89
-Revises: 1ef399cbd2b
-Create Date: 2014-06-23 19:59:59.004147
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '4f528e69e89'
-down_revision = '1ef399cbd2b'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
