--- alembic/versions/50b5a95300c_add_ownership_to_mod_lists.py
@@ -1,26 +1,0 @@
-"""Add ownership to mod lists
-
-Revision ID: 50b5a95300c
-Revises: 566d08c75c6
-Create Date: 2014-09-27 15:03:52.555439
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '50b5a95300c'
-down_revision = '566d08c75c6'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('modlist', sa.Column('user_id', sa.Integer(), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('modlist', 'user_id')
-    ### end Alembic commands ###
