--- alembic/versions/5330eb08753_add_featured_mods.py
@@ -1,22 +1,0 @@
-"""Add featured mods
-
-Revision ID: 5330eb08753
-Revises: 4f528e69e89
-Create Date: 2014-06-24 15:59:03.654017
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '5330eb08753'
-down_revision = '4f528e69e89'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
