--- alembic/versions/59abfc7b45d_add_download_path_to_modrevision.py
@@ -1,26 +1,0 @@
-"""Add download_path to ModRevision
-
-Revision ID: 59abfc7b45d
-Revises: 4ea6bc2045a
-Create Date: 2014-06-10 01:45:23.302614
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '59abfc7b45d'
-down_revision = '4ea6bc2045a'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('modversion', sa.Column('download_path', sa.String(length=512), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('modversion', 'download_path')
-    ### end Alembic commands ###
