--- alembic/versions/59c74736a7f_add_mod_follower_events.py
@@ -1,22 +1,0 @@
-"""Add mod follower events
-
-Revision ID: 59c74736a7f
-Revises: 3f35f3fa395
-Create Date: 2014-06-27 20:38:16.862316
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '59c74736a7f'
-down_revision = '3f35f3fa395'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
