--- alembic/versions/e591232705_add_download_events.py
@@ -1,22 +1,0 @@
-"""Add download events
-
-Revision ID: e591232705
-Revises: 263bcce1dc0
-Create Date: 2014-06-25 14:45:50.542411
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = 'e591232705'
-down_revision = '263bcce1dc0'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
