--- alembic/versions/e6fcc34862_expand_size_of_background_field.py
@@ -1,26 +1,0 @@
-"""Expand size of background field
-
-Revision ID: e6fcc34862
-Revises: 26f3fd0405c
-Create Date: 2015-02-08 14:29:38.200109
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = 'e6fcc34862'
-down_revision = '26f3fd0405c'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('mod', 'ckan_name')
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('mod', sa.Column('ckan_name', sa.VARCHAR(), autoincrement=False, nullable=True))
-    ### end Alembic commands ###
